+++ lib/tiny_rest_client/response.rb
@@ -6,14 +6,14 @@ class Response
   attr_reader :code, :status, :body, :headers
 
   def initialize(code: nil, status: nil, body: nil, headers: nil)
-    @code = code
+    @code = code&.to_i
     @status = status
     @headers = headers
     @body = parse_body(body)
   end
 
   def success?
-    code.to_i.between?(200, 299)
+    code.between?(200, 299)
   end
 
   def failure?